refactor(sdk): extract batch operations from LifecycleManager
Move the batch lifecycle methods (batchCreateAssets, batchPublishToWeb,
batchInscribeOnBitcoin + its single/individual-transaction helpers,
batchTransferOwnership, and the size/savings estimators) into a new
BatchLifecycleOperations collaborator. LifecycleManager keeps the same public
batch method signatures as thin delegators and passes itself as the core-ops
provider, so per-asset behavior and event emission are unchanged.

Pure move: LifecycleManager drops from 2178 to 1604 lines; all 410
lifecycle+integration tests pass identically (0 behavior change). tsc stays at
0 errors.
